Final Change to Project 5
I should have a push or two left for the audio of project 5, but the
code is complete. There's a regular Pokedex class and the Pokedex class
that includes audio.

I managed to fix a bug that would cause the audio description of the
Pokemon continue to play even when you picked another Pokemon.

I've included a .bat file that will batch convert the .mp4 files I am
given into .wav files for the program. It keeps track of what hasn't
been converted yet and will continue to convert the files as they are
given to me.